pyproject.toml
[project]
name = "telegram-bot-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "pymongo[srv]>=4.11.3",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.5",
    "python-telegram-bot>=22.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]

[tool.ruff]
line-length=79
